Commander subclass picker works in leaderbuilder
AWWW YEAHHHH!!!!!!!!

I had to double check a lot of trigger criteria.
the kicker was "admiral" in a bunch of them so I had to
swap in "commander" :) :) :)

I *still* don't know how it works that clicking a
subclass icon makes the subclas picker window "go away"
to reveal the trait picker window.

I suspect the dummy effect button somehow resets the
x/y of the window to 0,0