Collect subclasses into a list if there are multiple has_trait definitions
In the quest to convert traits file data to something a modern
programming language can work with, because this mod needs to
make lots of tooltips, converting to standard YAML has been tough.

Because PDX script can contain duplicate keys, which are not allowed
in standard YAML, nor can really be considered valid JSON
for that matter.

So this workaround finds multiple has_trait = subclass_*
and puts all the values into a single list