Finish implementing change_species_characteristics
I tested this over modify_species which the effects files say
creates a new species, but change characteristics doesnt create
a new species.

The old way was working but probably because of unintended side effects.
this way is more stable.

tested with MAMP, Bio Mech (new species and change primary)
the species properly inherit habitability characteristcs of the
parent species